refactor (#14 - 2): components relocated to their directories
