--- frontend/forcedepartment-app/src/components/About.js
@@ -1,16 +1,0 @@
-import {useEffect} from "react";
-
-const About = () => {
-
-    useEffect(()=> {
-        document.title = "Special Department | About us";
-    },[]);
-
-    return (
-        <div>
-            <p>This is the special department</p>
-        </div>
-    )
-}
-
-export default About
