--- frontend/forcedepartment-app/src/components/Footer.js
@@ -1,15 +1,0 @@
-import './Footer.css';
-
-const Footer = () => {
-    return (
-        <div id="footer">
-            <hr />
-            <div className="copyright">
-                <p className="footer-text">Special Department ©</p>
-            </div>
-        </div>
-    )
-}
-
-
-export default Footer
