--- frontend/forcedepartment-app/src/components/Header.css
@@ -1,205 +1,0 @@
-*{
-    padding: 0;
-    margin: 0;
-
-}
-
-body{
-    background: url('https://awsinsurance.com/wp-content/uploads/2015/11/background-clean.jpg') no-repeat center;
-    background-size: cover;
-    width: 100%;
-    height: 100%;
-    box-sizing: border-box;
-    font-family: sans-serif;
-}
-
-hr {
-    color: #a9a9a9;
-    opacity: 0.3;
-}
-
-#specificWorkerText {
-    font-size: 26px;
-    font-weight: 550;
-    text-align: center;
-    margin-top: 30px;
-}
-
-.nav{
-    display: flex;
-    justify-content: space-between;
-    align-content: center;
-    background: #000;
-    color: #fff;
-    padding: 1em 2em;
-}
-
-.nav ul{
-    display: flex;
-    justify-content: space-between;
-    align-content: center;
-}
-
-.navbar {
-    float: left;
-    font-size: 1.5em;
-}
-
-.nav li{
-    list-style: none;
-}
-
-.nav button {
-    background: none;
-    border: none;
-    color: #fff;
-    padding-right: 2em;
-    text-transform: uppercase;
-    font-size: 1em;
-}
-
-.nav button:hover {
-    cursor:pointer;
-}
-
-.nav a{
-    text-decoration: none;
-    color: #fff;
-    padding-right: 2em;
-    text-transform: uppercase;
-    font-size: 1em;
-}
-
-.nav a:hover, .nav button:hover{
-    color: rgb(216, 185, 10);
-}
-
-.logo a{
-    text-decoration: none;
-    color: #fff;
-    font-size: 1.5em;
-    text-transform: uppercase;
-}
-
-.filter-sidenav {
-    height: 100%;
-    width: 185px;
-    z-index: 1;
-    top: 40px;
-    left: 0;
-    background-color: #000;
-    overflow-x: hidden;
-    padding-top: 150px;
-}
-
-.filter-sidenav p {
-    margin-bottom: 55px;
-    color: #fff;
-}
-
-.filter-sidenav input {
-    padding: 6px 15px 6px 16px;
-    text-decoration: none;
-    font-size: 21px;
-    display: block;
-}
-
-.filter-sidenav input::placeholder {
-    padding-left: 20px;
-}
-
-.filter-sidenav input:focus-within {
-    color: #000;
-    background-image:none;
-}
-
-.filter-sidenav input::placeholder:focus-within {
-    padding-left: 0px;
-}
-
-.filter-sidenav button {
-    padding: 6px;
-}
-
-.filter-sidenav button:hover {
-}
-
-
-.filter-sidenav img{
-    padding-left:15px;
-    float: left;
-    width: 32px;
-    height: 32px;
-}
-
-.searchbar {
-    display: block;
-}
-
-.worker-card {
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.4);
-    transition: 0.3s;
-    height: 435px;
-    width: 315px;
-    display: inline-grid;
-    text-align: center;
-    margin-left: 45px;
-    margin-top: 20px;
-    border-radius: 3px;
-    background-color: rgba(255, 255, 255, 0.8);
-}
-
-.worker-detail img {
-    margin-top: 15px;
-    margin-bottom: 15px;
-    height: 280px;
-    width: 250px;
-}
-
-.worker-detail a {
-    text-decoration: none;
-    font-weight: 550;
-    color: #000;
-}
-
-.worker-card:hover {
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-}
-
-.web-information {
-    text-align: center;
-    margin-top: 30px;
-    padding: 10px;
-}
-
-.worker-card h3 {
-    text-align: center;
-}
-
-
-
-.copyright {
-    margin-top: 30px;
-    font-size: 17px;
-    font-weight: bold;
-}
-
-
-
-
-.main {
-    margin-left: 160px; /* Same as the width of the sidenav */
-    font-size: 28px; /* Increased text to enable scrolling */
-    padding: 0px 10px;
-}
-
-@media screen and (max-height: 450px) {
-    .sidenav {padding-top: 15px;}
-    .sidenav a {font-size: 18px;}
-}
-
-
-.worker-card-link {
-    text-decoration: none;
-    color: black;
-}
