--- frontend/forcedepartment-app/src/components/MenuPoint.js
@@ -1,21 +1,0 @@
-
-const MenuPoint = (props) => {
-    return (
-        <>
-        <div className="mainMenuPoint" onClick={props.listMenuPoint}>{props.menuTitle}</div>
-        {props.showMenuPoint && 
-            <ul>
-                {props.allMenuPoints.map((menuPoint) => (
-                    <li 
-                        onClick={() => {props.sideBarHandler(props.fetchRoutePart, menuPoint)}} 
-                        className="menuPoint">
-                        {menuPoint}
-                    </li>))}
-            </ul>
-        }
-        </>
-    )
-}
-
-
-export default MenuPoint
