--- frontend/forcedepartment-app/src/components/Profile.css
@@ -1,163 +1,0 @@
-*{
-    padding: 0;
-    margin: 0;
-}
-
-body{
-    background: url('https://awsinsurance.com/wp-content/uploads/2015/11/background-clean.jpg') no-repeat center;
-    background-size: cover;
-    width: 100%;
-    height: 100%;
-    box-sizing: border-box;
-    font-family: sans-serif;
-}
-
-hr {
-    color: #a9a9a9;
-    opacity: 0.3;
-}
-
-.nav{
-    display: flex;
-    justify-content: space-between;
-    align-content: center;
-    background: #000;
-    color: #fff;
-    padding: 1em 6em;
-}
-
-.nav ul{
-    display: flex;
-    justify-content: space-between;
-    align-content: center;
-}
-
-.navbar {
-    float: left;
-}
-
-.nav li{
-    list-style: none;
-}
-
-.nav button {
-    background: none;
-    border: none;
-    color: #fff;
-    padding-right: 2em;
-    text-transform: uppercase;
-    font-size: 1em;
-}
-
-.nav button:hover {
-    cursor:pointer;
-}
-
-.nav a{
-    text-decoration: none;
-    color: #fff;
-    padding-right: 2em;
-    text-transform: uppercase;
-    margin-top: 1%;
-    font-size: 21px;
-}
-
-.nav a:hover, .nav button:hover{
-    color: rgb(216, 185, 10);
-}
-
-.logo a{
-    text-decoration: none;
-    color: #fff;
-    font-size: 1.5em;
-    text-transform: uppercase;
-}
-
-.user-personal img{
-    width: 470px;
-    height: 540px;
-}
-
-.user-container {
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.4);
-    transition: 0.3s;
-    height: 800px;
-    width: 45%;
-    display: inline-block;
-    text-align: center;
-    margin-left: 45px;
-    margin-top: 60px;
-    border-radius: 3px;
-    background-color: rgba(255, 255, 255, 0.8);
-}
-
-.user-personal h1{
-    margin-top: 30px;
-}
-
-.user-personal p{
-    margin-top: 20px;
-    font-size: 21px;
-}
-
-#h2-profession{
-    margin-top: 115px;
-}
-
-.user-description h2{
-    margin-top: 40px;
-}
-
-.user-personal h2 {
-    margin-top: 70px;
-}
-
-.professions {
-    font-size: 20px;
-}
-
-.user-description p {
-    margin-top: 30px;
-    font-size: 22px;
-}
-
-.user-description-container {
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.4);
-    transition: 0.3s;
-    height: 500px;
-    width: 47%;
-    float: right;
-    vertical-align: middle;
-    display: inline-block;
-    text-align: center;
-    margin-right: 30px;
-    margin-top: 220px;
-    border-radius: 3px;
-    background-color: rgba(255, 255, 255, 0.8);
-}
-
-.user-profile-container {
-    float: left;
-}
-
-.user-description-content {
-    padding-left: 40px;
-    padding-right: 40px;
-}
-
-.footer{
-    margin-top: 90% !important;
-    color: rgb(255, 255, 255);
-    text-align: center;
-    background-color: black;
-}
-
-.footer-text {
-    padding-bottom: 30px;
-}
-
-.copyright {
-    margin-top: 30px;
-    font-size: 17px;
-    font-weight: bold;
-}
