--- frontend/forcedepartment-app/src/components/Sidebar.css
@@ -1,35 +1,0 @@
-.sidenav {
-    width: 260px;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    overflow-x: hidden;
-    padding-top: 20px;
-    margin-top: 20px;
-    float: left;
-    box-shadow: 5px 5px 8px #999999;
-    height: auto;
-    background-color: rgba(10, 10, 10, 0.8);
-    margin-left: 8%;
-}
-
-.mainMenuPoint {
-    cursor: pointer;
-    padding: 12px 8px 12px 16px;
-    font-size: 25px;
-    color: #babdc0;
-    display: block;
-}
-
-.menuPoint:hover, .mainMenuPoint:hover {
-    color: white;
-    
-}
-
-.menuPoint {
-    cursor: pointer;
-    padding: 2px 8px 2px 22px;
-    font-size: 18px;
-    color: #babdc0;
-    display: block;
-}
