--- frontend/forcedepartment-app/src/components/css/Login.css
@@ -1,107 +1,0 @@
-body {
-    background: url('https://awsinsurance.com/wp-content/uploads/2015/11/background-clean.jpg') no-repeat fixed center center;
-    background-size: cover;
-    font-family: Montserrat;
-}
-
-#errorMessage {
-    color:darkred;
-    font-size: 18px;
-    text-align:center;
-}
-
-.login-block {
-    width: 350px;
-    padding: 20px;
-    background-color: rgba(255, 255, 255, 1);
-    border-radius: 5px;
-    border: 15px solid #847a65;
-    margin: 240px auto;
-}
-
-.mainPageHref {
-    margin: 16px;
-}
-
-.mainPageHref a{
-    text-decoration: none;
-    color: #fff;
-}
-
-.login-block h1 {
-    text-align: center;
-    color: #000;
-    font-size: 18px;
-    font-family: Montserrat;
-    text-transform: uppercase;
-    margin-top: 0;
-    margin-bottom: 10px;
-}
-
-.login-block input {
-    width: 100%;
-    height: 42px;
-    box-sizing: border-box;
-    border-radius: 5px;
-    border: 1px solid #ccc;
-    margin-bottom: 20px;
-    font-size: 14px;
-    font-family: Montserrat;
-    padding: 0 20px 0 50px;
-    outline: none;
-}
-
-.login-block input#email {
-    background: #fff url('http://i.imgur.com/u0XmBmv.png') 20px top no-repeat;
-    background-size: 16px 80px;
-}
-
-.login-block input#email:focus {
-    background: #fff url('http://i.imgur.com/u0XmBmv.png') 20px bottom no-repeat;
-    background-size: 16px 80px;
-}
-
-.login-block input#password {
-    background: #fff url('http://i.imgur.com/Qf83FTt.png') 20px top no-repeat;
-    background-size: 16px 80px;
-}
-
-.login-block input#password:focus {
-    background: #fff url('http://i.imgur.com/Qf83FTt.png') 20px bottom no-repeat;
-    background-size: 16px 80px;
-}
-
-.login-block input:active, .login-block input:focus {
-    border: 1px solid #847a65;
-}
-
-.login-block button {
-    width: 100%;
-    height: 40px;
-    background: #847a65;
-    box-sizing: border-box;
-    border-radius: 5px;
-    border: 1px solid #9d7a65;
-    color: #fff;
-    font-weight: bold;
-    text-transform: uppercase;
-    font-size: 14px;
-    font-family: Montserrat;
-    outline: none;
-    cursor: pointer;
-}
-
-.login-block button:hover {
-    background: #737a65;
-}
-
-.mainPageHref p {
-    margin-top: 25px;
-    font-size: 17px;
-    color: dimgray;
-}
-
-#wrongLogin {
-    color:darkred;
-    text-align:center;
-}
