--- frontend/forcedepartment-app/src/components/css/Register.css
@@ -1,255 +1,0 @@
-body {
-    background: url('https://awsinsurance.com/wp-content/uploads/2015/11/background-clean.jpg') no-repeat fixed center center;
-    background-size: cover;
-    width: 100%;
-    height: 100%;
-
-}
-
-body,
-input,
-textarea,
-select {
- 
-    font-size: 16px;
-    color: #4c4c4c;
-}
-
-textarea {
-    width: 275px;
-    height: 150px;
-}
-
-.register-form {
-    text-align: center;
-    margin: 40px auto;
-    width: 343px;
-    height: 687px;
-    -webkit-border-radius: 8px/7px;
-    -moz-border-radius: 8px/7px;
-    border-radius: 8px/7px;
-    background-color: #ebebeb;
-    -webkit-box-shadow: 1px 2px 5px rgba(0, 0, 0, .31);
-    -moz-box-shadow: 1px 2px 5px rgba(0, 0, 0, .31);
-    box-shadow: 1px 2px 5px rgba(0, 0, 0, .31);
-    border: solid 1px #cbc9c9;
-}
-
-form {
-    margin-top: 3%;
-}
-
-h1 {
-    font-size: 32px;
-    font-weight: 300;
-    color: #4c4c4c;
-    text-align: center;
-    padding-top: 10px;
-    margin-bottom: 10px;
-}
-
-#wrongLogin p {
-    font-size: 16px;
-    font-weight: 300;
-    color: #4c4c4c;
-    text-align: center;
-    padding-top: 10px;
-
-}
-
-
-hr {
-    color: #a9a9a9;
-    opacity: 0.3;
-}
-
-input[type=number] {
-    width: 100px;
-    height: 21px;
-    -webkit-border-radius: 0px 4px 4px 0px/5px 5px 4px 4px;
-    -moz-border-radius: 0px 4px 4px 0px/0px 0px 4px 4px;
-    border-radius: 0px 4px 4px 0px/5px 5px 4px 4px;
-    background-color: #fff;
-    -webkit-box-shadow: 1px 2px 5px rgba(0, 0, 0, .09);
-    -moz-box-shadow: 1px 2px 5px rgba(0, 0, 0, .09);
-    box-shadow: 1px 2px 5px rgba(0, 0, 0, .09);
-    border: solid 1px #cbc9c9;
-    margin-left: -5px;
-    margin-top: 13px;
-    padding-left: 10px;
-}
-
-.userType {
-    margin-left: 8px;
-    margin-top: 20px;
-    text-align: center;
-}
-
-input[type=text], input[type=password], input[type=email], input[type=date] {
-    width: 200px;
-    height: 39px;
-    -webkit-border-radius: 0px 4px 4px 0px/5px 5px 4px 4px;
-    -moz-border-radius: 0px 4px 4px 0px/0px 0px 4px 4px;
-    border-radius: 0px 4px 4px 0px/5px 5px 4px 4px;
-    background-color: #fff;
-    -webkit-box-shadow: 1px 2px 5px rgba(0, 0, 0, .09);
-    -moz-box-shadow: 1px 2px 5px rgba(0, 0, 0, .09);
-    box-shadow: 1px 2px 5px rgba(0, 0, 0, .09);
-    border: solid 1px #cbc9c9;
-    margin-left: -5px;
-    margin-top: 13px;
-    padding-left: 10px;
-}
-
-input {
-    padding-top: 6px;
-}
-
-input[type=password] {
-}
-
-.userType {
-    display: inline-block;
-    cursor: pointer;
-    font-weight: 500;
-    position: relative;
-    overflow: hidden;
-    text-align: center;
-    font-size: 19px;
-    margin-left: 20px;
-    margin-right: 23px;
-    margin-top: 35px;
-}
-
-#wrongLogin {
-    color:darkred;
-    font-size: 18px;
-    text-align:center;
-    margin-bottom: 15px;
-}
-
-.register-panel {
-    font-family: Montserrat;
-}
-
-.register-panel button {
-    font-size: 14px;
-    font-weight: 600;
-    color: white;
-    padding: 6px 25px 0px 20px;
-    margin: 10px 30px 20px 0px;
-    display: inline-block;
-    float: right;
-    text-decoration: none;
-    width: 120px;
-    height: 40px;
-    -webkit-border-radius: 5px;
-    -moz-border-radius: 5px;
-    border-radius: 5px;
-    background-color: steelblue;
-    -webkit-box-shadow: 0 3px rgba(58, 87, 175, .75);
-    -moz-box-shadow: 0 3px rgba(58, 87, 175, .75);
-    box-shadow: 0 3px rgba(58, 87, 175, .75);
-    transition: all 0.1s linear 0s;
-    top: 0px;
-    position: relative;
-    margin-top: 50px;
-}
-
-.register-panel button:hover {
-    top: 3px;
-    background-color: #2e458b;
-    -webkit-box-shadow: none;
-    -moz-box-shadow: none;
-    box-shadow: none;
-}
-
-input[type="date"]::-webkit-clear-button {
-    display: none;
-}
-
-input[type="date"]::-webkit-inner-spin-button {
-    display: none;
-}
-
-input[type="date"]::-webkit-calendar-picker-indicator {
-    color: #2c3e50;
-}
-
-input[type="date"]:focus {
-    color: #95a5a6;
-    box-shadow: none;
-    -webkit-box-shadow: none;
-    -moz-box-shadow: none;
-}
-
-#addProfession {
-    width: 50px;
-    height: 22px;
-    margin-left: 20px;
-    color: white;
-    border-radius: 5px;
-    background-color: steelblue;
-    transition: all 0.1s linear 0s;
-    font-size: 14px;
-    font-weight: bold;
-}
-
-#addProfession:hover {
-    background-color: #2e458b;
-}
-
-#removeProfession {
-    width: 20px;
-    height: 20px;
-}
-
-.register-worker-panel textarea, input, h2, select{
-    margin-top: 20px;
-}
-
-.register-worker-panel li {
-    list-style-type: none;
-    display: inline-block;
-    margin-top: 10px;
-    margin-right: 10px;
-    color: indianred;
-    font-weight: bold;
-    font-size: 16px;
-}
-
-.register-worker-panel #actionButtons {
-    font-size: 14px;
-    font-weight: 600;
-    color: white;
-    padding: 6px 25px 0px 20px;
-    margin: 10px 30px 20px 0px;
-    display: inline-block;
-    float: right;
-    text-decoration: none;
-    width: 120px;
-    height: 40px;
-    -webkit-border-radius: 5px;
-    -moz-border-radius: 5px;
-    border-radius: 5px;
-    background-color: steelblue;
-    -webkit-box-shadow: 0 3px rgba(58, 87, 175, .75);
-    -moz-box-shadow: 0 3px rgba(58, 87, 175, .75);
-    box-shadow: 0 3px rgba(58, 87, 175, .75);
-    transition: all 0.1s linear 0s;
-    top: 0px;
-    position: relative;
-    margin-top: 50px;
-}
-
-.register-worker-panel #actionButtons:hover {
-    top: 3px;
-    background-color: #2e458b;
-    -webkit-box-shadow: none;
-    -moz-box-shadow: none;
-    box-shadow: none;
-}
-
-select {
-    width: 50%;
-}
